git redo, see if this fixes anything

+++ scripts/buildArch.sh
@@ -1,21 +1,31 @@
-#! /bin/bash
+#Temp code to make sure we only work on armhf for now
+if [ "$1" != "armhf" ]; then
+        echo "only armhf is supported right now"
+        exit
+fi
 
+export POPNAME=archlinxarm
 export ARCH_DIR=output/${1}
 export ROOTFS_DIR=$ARCH_DIR/rootfs
+
+#Current workaround for mounting issues with chroot
+export CHROOTCMD="proot -0 -b /run -b /sys -b /dev -b /proc -b /mnt --rootfs=$ROOTFS_DIR"
 
 rm -rf $ARCH_DIR
 mkdir -p $ARCH_DIR
 rm -rf $ROOTFS_DIR
 mkdir -p $ROOTFS_DIR
 
-# Here we untar the ARM port, need to figure out what to do with the x86 version
+# wget http://fl.us.mirror.archlinuxarm.org/os/ArchLinuxARM-armv7-latest.tar.gz
+tar -zxvf ArchLinuxARM-armv7-latest.tar.gz $ROOTFS_DIR
 
-wget http://fl.us.mirror.archlinuxarm.org/os/ArchLinuxARM-armv7-latest.tar.gz
-tar -zxvf ArchLinuxARM-armv5-latest.tar.gz $ROOTFS_DIR
-
-# The following just sets up a few files for UserLAnd's benefit, mostly nameserver and path stuff
+# first setup the chroot environment
+cp /usr/bin/qemu-arm-static $ROOTFS_DIR/usr/bin
+chroot ./$ROOTFS_DIR /bin/bash
+# let's see if the above works..
 
 echo "127.0.0.1 localhost" > $ROOTFS_DIR/etc/hosts
+rm $ROOTFS_DIR/etc/resolv.conf
 echo "nameserver 8.8.8.8" > $ROOTFS_DIR/etc/resolv.conf
 echo "nameserver 8.8.4.4" >> $ROOTFS_DIR/etc/resolv.conf
 
@@ -25,25 +35,24 @@ echo "unset LD_LIBRARY_PATH" >> $ROOTFS_DIR/etc/profile.d/userland.sh
 echo "export LIBGL_ALWAYS_SOFTWARE=1" >> $ROOTFS_DIR/etc/profile.d/userland.sh
 chmod +x $ROOTFS_DIR/etc/profile.d/userland.sh
 
-pacman-key init
-pacman-key --populate archlinux
-
-# Copy the scripts and tar up everything
-
 cp scripts/addNonRootUser.sh $ROOTFS_DIR
 chmod 777 $ROOTFS_DIR/addNonRootUser.sh
-DEBIAN_FRONTEND=noninteractive DEBCONF_NONINTERACTIVE_SEEN=true LC_ALL=C LANGUAGE=C LANG=C chroot $ROOTFS_DIR ./addNonRootUser.sh
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD ./addNonRootUser.sh
 rm $ROOTFS_DIR/addNonRootUser.sh
 
-cp scripts/shrinkRootfs.sh $ROOTFS_DIR
-chmod 777 $ROOTFS_DIR/shrinkRootfs.sh
-DEBIAN_FRONTEND=noninteractive DEBCONF_NONINTERACTIVE_SEEN=true LC_ALL=C LANGUAGE=C LANG=C chroot $ROOTFS_DIR ./shrinkRootfs.sh
-rm $ROOTFS_DIR/shrinkRootfs.sh
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman-key --init
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman-key --populate $POPNAME
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman-key -Syy --noconfirm
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman-key -Su --noconfirm
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman -S pacman-contrib base base-devel sudo tigervnc xterm xorg-twm expect --noconfirm
 
-tar --exclude='dev/*' -czvf $ARCH_DIR/rootfs.tar.gz -C $ROOTFS_DIR .
+tar --exclude='dev/*' -czvf $ARCH_DIR/rootfs.tar.gz -C $ROOTFS_DIR
 
 #build disableselinux to go with this release
 cp scripts/disableselinux.c $ROOTFS_DIR
-DEBIAN_FRONTEND=noninteractive DEBCONF_NONINTERACTIVE_SEEN=true \
- LC_ALL=C LANGUAGE=C LANG=C chroot $ROOTFS_DIR gcc -shared -fpic disableselinux.c -o libdisableselinux.so
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD gcc -shared -fpic disableselinux.c -o libdisableselinux.so 
 cp $ROOTFS_DIR/libdisableselinux.so $ARCH_DIR/libdisableselinux.so
+
+#get busybox to go with the release
+LC_ALL=C LANGUAGE=C LANG=C $CHROOTCMD pacman -S busybox --noconfirm
+cp $ROOTFS_DIR/bin/busybox $ARCH_DIR/busybox
